backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
nibabel
pillow
